Updated tests and dependencies
* Since I'm using the cloudera CDH 5.9.1 I added repo and versions according to that repo
  else Zookeeper discovery wouldn't work.
* Added property in gradle.properties to specify distributionRepo.
