+++ docker/build.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+docker build -t hive-kerberos:latest .
