+++ docker/zookeeper/jaas.conf
@@ -1,0 +1,16 @@
+Server {
+  com.sun.security.auth.module.Krb5LoginModule required
+  useKeyTab=true
+  keyTab="/etc/zookeeper/conf/zookeeper.keytab"
+  storeKey=true
+  useTicketCache=false
+  principal="zookeeper/hadoop-master@LABS.TERADATA.COM";
+};
+Client {
+  com.sun.security.auth.module.Krb5LoginModule required
+  useKeyTab=true
+  keyTab="/etc/zookeeper/conf/zkcli.keytab"
+  storeKey=true
+  useTicketCache=false
+  principal="zkcli@LABS.TERADATA.COM";
+};
