+++ gradle.properties
@@ -1,10 +1,11 @@
 #Project
-version=0.0.2
+version=0.0.3
 gradleWrapperVersion=4.4.1
 
 #Dependencies
-hiveVersion=1.1.0
-hadoopVersion=2.6.0
+distributionRepo=https://repository.cloudera.com/content/repositories/releases/
+hiveVersion=1.1.0-cdh5.9.1
+hadoopVersion=2.6.0-cdh5.9.1
 
 #Test
 assertjVersion=3.8.0
